FInalProject/requirements.txt
numpy
torch>=1.11
torchvision
Pillow
random
tqdm
os
pickle
math
opencv_python
matplotlib



